pyproject.toml
[tool.poetry]
name = "rag-experiments"
version = "0.1.0"
description = "RAG (Retrieval-Augmented Generation) experiments with PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
alembic = "^1.13.0"
requests = "^2.32.5"
tqdm = "^4.67.1"

[tool.poetry.scripts]
download-wikipedia = "scripts.download_wikipedia:main"
download-arxiv = "scripts.download_arxiv:main"
download-books = "scripts.download_books:main"
download-news = "scripts.download_news:main"
compose-up = "scripts.compose_up:main"
compose-down = "scripts.compose_down:main"
setup = "scripts.setup:main"
generate-questions = "scripts.generate_questions:main"
ingest-wikipedia = "scripts.ingest_wikipedia:main"
evaluate-retrieval = "scripts.evaluate_retrieval:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.1"
ipython = "^8.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
